% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcr_graph.R
\docType{data}
\name{GPCR_graph}
\alias{GPCR_graph}
\title{Signaling by GPCR}
\format{A graph object of 32 verticies and 173 edges:
\describe{
  \item{V}{gene symbol (human)}
  \item{E}{directed relationship for pathway}
}}
\source{
PathwayCommons \url{"https://reactome.org/content/detail/R-HSA-372790"}
}
\usage{
GPCR_graph
}
\description{
Reactome pathway R-HSA-372790 for the interactions in the signalling by G protein-coupled receptors (GPCRs)
}
\keyword{datasets}
