% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adjmatrix_graph.R
\name{Generate Adjacency Matrix from Graph Structure}
\alias{Generate Adjacency Matrix from Graph Structure}
\alias{make_adjmatrix.igraph}
\title{Generate Adjacency Matrix}
\usage{
make_adjmatrix.igraph(graph, directed = F)
}
\arguments{
\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}

\item{directed}{logical. Whether directed information is passed to the adjacency matrix.}
}
\description{
Compute the adjacency matrix of a (directed) \code{\link[igraph]{igraph}} structure, preserving node/column/row names (and direction).
}
\keyword{adjacency}
\keyword{graph}
\keyword{igraph}
\keyword{network}

