% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_commonlink_graph.R
\name{Generate Common Link Matrix from Graph Structure}
\alias{Generate Common Link Matrix from Graph Structure}
\alias{make_commonlink_adjmat}
\alias{make_commonlink_graph}
\title{Generate Common Link Matrix}
\usage{
make_commonlink_adjmat(adj_mat)

make_commonlink_graph(graph, directed = F)
}
\arguments{
\item{adj_mat}{precomputed adjacency matrix.}

\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}

\item{directed}{logical. Whether directed information is passed to the adjacency matrix.}
}
\description{
Compute the common link matrix of a (directed) \code{\link[igraph]{igraph}} structure, preserving node/column/row names (and direction).
}
\keyword{graph}
\keyword{igraph}
\keyword{neighbohood}
\keyword{network}
