% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_distance_graph.R
\name{make_distance_graph}
\alias{make_distance_graph}
\title{Generate Distance Matrix}
\usage{
make_distance_graph(g, directed = T, absolute = F)
}
\arguments{
\item{directed}{logical. Whether directed information is passed to the distance matrix.}

\item{absolute}{logical. Whether distances are scaled as the absolute difference from the diameter (maximum possible). Defaults to TRUE. The alternative is to calculate a relative difference from the diameter for a geometric decay in distance.}

\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}
}
\description{
Compute the distance matrix of using shortest paths of a (directed) \code{\link[igraph]{igraph}} structure, normalising by the diameter of the network, preserving node/column/row names (and direction).
}
\keyword{adjacency}
\keyword{graph}
\keyword{igraph}
\keyword{network}

