% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sigma_graph.R
\name{make_sigma_mat.matrix}
\alias{make_sigma_mat.igraph}
\alias{make_sigma_mat.matrix}
\alias{make_sigma_mat_dist.igraph}
\alias{make_sigma_mat_dist.matrix}
\title{Generate Sigma Matrix}
\usage{
make_sigma_mat.matrix(mat, cor)

make_sigma_mat.igraph(graph, cor, comm = F, directed = F)

make_sigma_mat_dist.matrix(mat, cor, absolute = F)

make_sigma_mat_dist.igraph(graph, cor, absolute = F)
}
\arguments{
\item{mat}{precomputed adjacency, commonlink, or scaled distance matrix.}

\item{cor}{numeric. Simulated maximum correlation/covariance of two adjacent nodes. Default to 0.8.}

\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}

\item{comm}{logical whether a common link matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).}

\item{directed}{logical. Whether directed information is passed to the distance matrix.}

\item{absolute}{logical. Whether distances are scaled as the absolute difference from the diameter (maximum possible). Defaults to TRUE. The alternative is to calculate a relative difference from the diameter for a geometric decay in distance.}
}
\description{
Compute the Sigma matrix from an \code{\link[igraph]{igraph}} structure or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}} and \code{\link[graphsim]{generate_expression}}.
}
\keyword{graph}
\keyword{igraph}
\keyword{mvtnorm}
\keyword{network}

