% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_state_matrix.R
\name{Make State Matrix}
\alias{Make State Matrix}
\alias{make_state_matrix}
\title{Make State Matrix}
\usage{
make_state_matrix(graph, state)
}
\arguments{
\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted as long as a shortest path can be computed.}

\item{state}{numeric vector. Vector of length E(graph). Sign used to calculate state matrix, may be an integer state or inferred directly from expect correlations for each edge.  may also be entered as text for "activating" or "inhibiting". May be applied a scalar across all edges or as a vector for each edge respectively.}
}
\description{
Functions to compute the matrix of states (1 for activating and -1 for inhibiting) for link signed correlations, from a vector of edge states to a signed adjacency matrix for use in \code{\link[graphsim]{generate_expression}}.
}
\keyword{graph}
\keyword{igraph}
\keyword{mvtnorm}
\keyword{network}
\keyword{simulation}

