% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi3k_akt_graph.R
\docType{data}
\name{Pi3K_AKT_graph}
\alias{Pi3K_AKT_graph}
\title{PI3K/AKT activation}
\format{A graph object of 275 vertices and 21106 edges:
\describe{
  \item{V}{gene symbol (human)}
  \item{E}{directed relationship for pathway}
  \item{state}{type of relationship (activating or inhibiting) as edge attribute}
}}
\source{
PathwayCommons \url{https://reactome.org/content/detail/R-HSA-198203}
}
\usage{
Pi3K_AKT_graph
}
\description{
Reactome pathway R-HSA-198203 for the interactions in the phosphoinositide-3-kinase activation of Protein kinase B (PKB), also known as Akt
}
\keyword{datasets}
