% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi3k_graph.R
\docType{data}
\name{Pi3K_graph}
\alias{Pi3K_graph}
\title{PI3K Cascade}
\format{
A graph object of 35 vertices and 251 edges:
\describe{
  \item{V}{gene symbol (human)}
  \item{E}{directed relationship for pathway}
  \item{state}{type of relationship (activating or inhibiting) as edge attribute}
}
}
\source{
PathwayCommons \url{https://reactome.org/content/detail/R-HSA-109704}
}
\usage{
Pi3K_graph
}
\description{
Reactome pathway R-HSA-109704 for the interactions in the phosphoinositide-3-kinase cascade
}
\keyword{datasets}
