% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsim-package.R
\docType{package}
\name{graphsim-package}
\alias{graphsim-package}
\alias{graphsim}
\title{The graphsim package}
\description{
graphsim is a package to simulate normalised expression data from networks 
for biological pathways using \sQuote{\code{igraph}} objects and multivariate
normal distributions.
}
\section{Introduction}{

This package enables the generation of simulated gene expression datasets 
containing pathway relationships from a known underlying network.
These simulated datasets can be used to evaluate various bioinformatics 
methodologies, including statistical and network inference procedures.

These are computed by 1) resolving inhibitory states to derive a consistent
matrix of positive and negative edges, 2) inferring relationships between
nodes from paths in the graph, 3) weighting these in a Sigma (\eqn{\Sigma}) 
covariance matrix and 4) using this to sample a multivariate normal 
distribution.
}

\section{Getting Started}{

lorem ipsum
}

\section{Creating Input Data}{

lorem ipsum

#' @section Internal Functions:
lorem ipsum
}

\section{Examining Step-by-Step}{

lorem ipsum
}

\section{Visualization}{

lorem ipsum
}

\section{Further information}{

  The graphsim pacakge is published in the \emph{Journal of Open Source
   Software}. See the paper here for more details:
   \href{https://doi.org/10.21105/joss.02161}{doi: 10.21105/joss.02161}
  
  The graphsim GitHub repository is here:
  \href{https://github.com/TomKellyGenetics/graphsim}{TomKellyGenetics/graphsim}
  You can find the development version and submit an
  \href{https://github.com/TomKellyGenetics/graphsim/issues/new/choose}{issue}
  if you have questions or comments.
}

