% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsim-package.R
\docType{package}
\name{graphsim-package}
\alias{graphsim-package}
\alias{graphsim}
\title{The graphsim package}
\description{
graphsim is a package to simulate normalised expression data from networks 
for biological pathways using \sQuote{\code{igraph}} objects and multivariate
normal distributions.
}
\details{
This package provides functions to develop simulated continuous data 
(e.g., gene expression) from a sigma covariance matrix derived from a 
graph structure in \sQuote{\code{igraph}} objects. Intended to extend 
\sQuote{\code{mvtnorm}} to take 'igraph' structures rather than sigma 
matrices as input. This allows the use of simulated data that correctly
accounts for pathway relationships and correlations. Here we present
a versatile statistical framework to simulate correlated gene expression
data from biological pathways, by sampling from a multivariate normal
distribution derived from a graph structure. This package allows the
simulation of biologicalpathways from a graph structure based on a
statistical model of gene expression, such as simulation of expression
profiles that of log-transformed and normalised data from microarray
and RNA-Seq data.
}
\section{Introduction}{

This package enables the generation of simulated gene expression datasets 
containing pathway relationships from a known underlying network.
These simulated datasets can be used to evaluate various bioinformatics 
methodologies, including statistical and network inference procedures.

These are computed by 1) resolving inhibitory states to derive a consistent
matrix of positive and negative edges, 2) inferring relationships between
nodes from paths in the graph, 3) weighting these in a Sigma (\eqn{\Sigma}) 
covariance matrix and 4) using this to sample a multivariate normal 
distribution.
}

\section{Getting Started}{

The \code{\link[graphsim]{generate_expression}} function is a wrapper 
around all necessary functions to give a final simulated dataset.

Here we set up an example graph object using the 
\code{\link[igraph:aaa-igraph-package]{igraph}} package.

\preformatted{
library("igraph")
graph_structure_edges <- rbind(c("A", "C"), c("B", "C"), c("C", "D"),c("D", "E"),
                               c("D", "F"), c("F", "G"), c("F", "I"), c("H", "I"))
graph_structure <- graph.edgelist(graph_structure_edges, directed = TRUE)
}

Then we can call \code{\link[graphsim]{generate_expression}} to return
the simulated data based on the relationships defined in the graph
structure. Various options are available to fine-tune this.

\preformatted{
expr <- generate_expression(100, graph_structure,
                            cor = 0.8,
                            mean = 0,
                            sd = 1,
                            comm = FALSE,
                            dist = TRUE,
                            absolute = FALSE,
                            laplacian = FALSE)
}

Here we can see the final result.

\preformatted{
dim(expr)

library("gplots")
heatmap.2(expr,
          scale = "none",
          trace = "none",
          col = bluered(50),
          colsep = 1:4,
          rowsep = 1:4)
}

This dataset consists of 9 rows (one for each vertex or gene)
in the graph and 100 columns (one for each sample or observation).

Input with an adjacency matrix is available using the
\code{\link[graphsim:generate_expression]{generate_expression_mat}}
function.
}

\section{Creating Input Data}{

lorem ipsum (igraph / matrix functions / import from pkg data)
}

\section{Internal Functions}{

lorem ipsum (sigma, dist, state)
}

\section{Examining Step-by-Step}{

lorem ipsum (plotting sigma and \code{cor(t(expr))})
}

\section{Visualization}{

lorem ipsum (plot_directed)
}

\section{Further information}{

  The graphsim package is published in the \emph{Journal of Open Source Software}.
  See the paper here for more details:
   \href{https://doi.org/10.21105/joss.02161}{doi: 10.21105/joss.02161}
  
  The graphsim GitHub repository is here:
  \href{https://github.com/TomKellyGenetics/graphsim}{TomKellyGenetics/graphsim}
  You can find the development version and submit an
  \href{https://github.com/TomKellyGenetics/graphsim/issues/new/choose}{issue}
  if you have questions or comments.
}

\section{Citation}{

  
  To cite package 'graphsim' in publications use:
  
  Kelly, S.T. and Black, M.A. (2020). graphsim: An R package for simulating gene
  expression data from graph structures of biological pathways.
  \emph{Journal of Open Source Software}, \bold{5}(51), 2161,
  \url{https://doi.org/10.21105/joss.02161}
  
  A BibTeX entry for LaTeX users is: \preformatted{
  @article{Kelly2020joss02161,
     doi = {10.21105/joss.02161},
     url ={https://doi.org/10.21105/joss.02161},
     year = {2020},
     publisher = {The Open Journal},
     volume = {5},
     number = {51},
     pages = {2161},
     author = {S. Thomas Kelly and Michael A. Black},
     title = {graphsim: An R package for simulating gene expression data from graph structures of biological pathways},
     journal = {Journal of Open Source Software} 
   }
 }
}

\seealso{
Publication at \emph{Journal of Open Source Software}:
 
 \itemize{
 \item \url{https://doi.org/10.21105/joss.02161}
 }
 
 GitHub repository:
 
 \itemize{
 \item \url{https://github.com/TomKellyGenetics/graphsim/}
 }
 
 Report bugs:
 
 \itemize{
 \item \url{https://github.com/TomKellyGenetics/graphsim/issues}
 }
 
 Contributions:
 
 \itemize{
 \item \url{https://github.com/TomKellyGenetics/graphsim/blob/master/CONTRIBUTING.md}
 }
}
\author{
\bold{Maintainer}:  Tom Kelly  \email{tom.kelly@riken.jp}

Authors:
\itemize{
\item Tom Kelly (RIKEN IMS) \href{https://orcid.org/0000-0003-3904-6690}{ORCID})
\item Mik Black (Otago University) (\href{0000-0003-1174-6054}{ORCID})
}

Reviewers:
\itemize{
\item Cory Brunson (UConn) (\href{0000-0003-3126-9494}{ORCID})
\item Mik Black (Otago University) (\href{0000-0003-1174-6054}{ORCID})
}

Editor: Mark Jensen (Frederick National Laboratory for Cancer Research)
}
