% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{make_sigma}
\alias{make_sigma}
\alias{make_sigma_mat_adjmat}
\alias{make_sigma_mat_comm}
\alias{make_sigma_mat_laplacian}
\alias{make_sigma_mat_graph}
\alias{make_sigma_mat_dist_adjmat}
\alias{make_sigma_mat_dist_graph}
\title{Generate Sigma Matrix}
\usage{
make_sigma_mat_adjmat(mat, cor = 0.8)

make_sigma_mat_comm(mat, cor = 0.8)

make_sigma_mat_laplacian(mat, cor = 0.8)

make_sigma_mat_graph(
  graph,
  cor = 0.8,
  comm = FALSE,
  laplacian = FALSE,
  directed = FALSE
)

make_sigma_mat_dist_adjmat(mat, state = NULL, cor = 0.8, absolute = FALSE)

make_sigma_mat_dist_graph(graph, state = NULL, cor = 0.8, absolute = FALSE)
}
\arguments{
\item{mat}{precomputed adjacency, laplacian, commonlink, or scaled distance matrix.}

\item{cor}{numeric. Simulated maximum correlation/covariance of two adjacent nodes. Default to 0.8.}

\item{graph}{An \code{\link[igraph]{igraph}} object. May be directed or weighted.}

\item{comm}{logical whether a common link matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).}

\item{laplacian}{logical whether a Laplacian matrix is used to compute sigma. Defaults to FALSE (adjacency matrix).}

\item{directed}{logical. Whether directed information is passed to the distance matrix.}

\item{state}{numeric vector. Vector of length E(graph). Sign used to calculate state matrix, may be an integer state or inferred directly from expected correlations for each edge. May be applied a scalar across all edges or as a vector for each edge respectively. May also be entered as text for "activating" or "inhibiting" or as integers for activating (0,1) or inhibiting (-1,2). Compatible with inputs for \code{\link[graphsim]{plot_directed}}. Also takes a pre-computed state matrix from \code{\link[graphsim]{make_state_matrix}} if applied to the same graph multiple times.}

\item{absolute}{logical. Whether distances are scaled as the absolute difference from
the diameter (maximum possible). Defaults to TRUE. The alternative is to calculate a
relative difference from the diameter for a geometric decay in distance.}
}
\value{
a numeric covariance matrix of values in the range [-1, 1]
}
\description{
Compute the Sigma matrix from an \code{\link[igraph]{igraph}} structure 
or pre-computed matrix. These are compatible with \code{\link[mvtnorm]{rmvnorm}} and
\code{\link[graphsim]{generate_expression}}.
}
\examples{

library("igraph")
graph_test_edges <- rbind(c("A", "B"), c("B", "C"), c("B", "D"))
graph_test <- graph.edgelist(graph_test_edges, directed = TRUE)
adjacency_matrix <- make_adjmatrix_graph(graph_test)
sigma_matrix <- make_sigma_mat_adjmat(adjacency_matrix, cor = 0.8)

}
\keyword{graph}
\keyword{igraph}
\keyword{mvtnorm}
\keyword{network}
